requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.38.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==23.2.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.51
botocore==1.35.51
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
cohere-aws==0.8.16
contourpy==1.3.0
courlan==1.3.2
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
databricks-sdk==0.36.0
dataclasses-json==0.6.7
datasets==3.0.1
dateparser==1.2.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphene==3.4.1
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
htmldate==1.9.3
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jusText==3.0.1
kiwisolver==1.4.7
langchain==0.3.3
langchain-aws==0.2.4
langchain-community==0.3.2
langchain-core==0.3.12
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.136
llama-index==0.10.22
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llamaindex-py-client==0.1.19
lxml==5.3.0
lxml_html_clean==0.4.1
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.17.2
mlflow-skinny==2.17.2
mock==4.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
olefile==0.47
openai==1.52.0
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pathos #==0.3.3
pillow==11.0.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypdf==4.3.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
ragas==0.1.0    # 0.2.0
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
sagemaker==2.232.3
sagemaker-core==1.0.13
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.0.6
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinysegmenter==0.3
tld==0.13
tldextract==5.1.3
tokenizers==0.20.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.5
trafilatura==2.0.0
transformers==4.45.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3
unstructured==0.16.0
unstructured-client #==0.26.1
urllib3==2.2.3
vertexai==1.71.1
Werkzeug==3.1.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.4
zipp==3.20.2